move QueryResult tests to their own module
